Remove version global Add some tests
