+++ src/components/Buttons/AutoRefreshSwitch.jsx
@@ -1,7 +1,7 @@
 import React, { PropTypes } from 'react';
 import cx from 'classnames';
 import store from '../../core/store';
-import { setAutoupdate } from '../../core/actions';
+import { setAutoupdate } from '../../core/legacy-actions';
 
 class AutoRefreshSwitch extends React.Component {
   static propTypes = {
