+++ templates/version.ejs
@@ -1,0 +1,5 @@
+{
+    "package": <%- version %>,
+    "git": "<%= git %>",
+    "debug": <%= debug %>
+}
